requirements.txt
# Core RAG stack
docling>=2.5.0
pymupdf>=1.24.0          # page-type detection (image-only vs digital) — fast pre-scan
qdrant-client>=1.9.0
sentence-transformers>=3.0.0
anthropic>=0.30.0

# MCP server
mcp>=1.0.0
uvicorn>=0.30.0
starlette>=0.37.0

# Document processing utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
tqdm>=4.66.0
rich>=13.0.0
loguru>=0.7.0
